added protection to the products routes

+++ backend/controllers/productsController.js
@@ -13,10 +13,11 @@ const Product = require('../models/productsModel')
 const getProducts = asyncHandler(async (req, res) => {
     const data = await Product.find()
     if (!data) {
-        res.status(500).json({ message: 'server error please try again' })
-        return
+        res.status(500)
+        throw new Error('server error please try again')
+    } else {
+        res.status(200).json(data)
     }
-    res.status(200).json(data)
 })
 
 // @desc    get one product
@@ -25,53 +26,53 @@ const getProducts = asyncHandler(async (req, res) => {
 const getProduct = asyncHandler(async (req, res) => {
     const data = await Product.findById(req.params.id)
     if (!data) {
-        res.status(500).json({ message: 'server error please try again' })
-        return
+        res.status(500)
+        throw new Error('server error please try again')
+    } else {
+        res.status(200).json(data)
     }
-    res.status(200).json(data)
 })
 
 // @desc    Add new product
 // @route   POST /api/products
 // @access  private
 const addProduct = asyncHandler(async (req, res) => {
-    const { category, images, age, pieces, isFeatured, features, highlights, details, name, price, brand, tags } = req.body
-    const newProduct = {
-        category,
-        images,
-        age,
-        pieces,
-        isFeatured,
-        features,
-        highlights,
-        details,
-        name,
-        price,
-        brand,
-        reviews: [],
-        tags,
-    }
+    if (req.user.type === 'Admin' && req.user.status === 'Active') {
+        const { category, images, age, pieces, isFeatured, features, highlights, details, name, price, brand, tags } = req.body
+        const newProduct = {
+            category,
+            images,
+            age,
+            pieces,
+            isFeatured,
+            features,
+            highlights,
+            details,
+            name,
+            price,
+            brand,
+            reviews: [],
+            tags,
+        }
 
-    // check for missing fields
-    const check = Object.keys(newProduct).filter(k => !newProduct[k])
-    if (check.length > 0) {
-        res.status(400).json({ message: 'missing fields', field: check })
-        return
-    }
-
-    // check if product exists
-    const exists = await Product.findOne({ name: newProduct.name })
-    if (exists) {
-        res.status(400).json({ message: 'already exists' })
-    }
-
-    // create the product
-    try {
-        const data = await Product.create(newProduct)
-        res.status(201).json(data)
-    } catch (err) {
+        // check if product exists
+        const exists = await Product.findOne({ name: newProduct.name })
+        if (exists) {
+            res.status(400)
+            throw new Error('already exists')
+        } else {
+            // create the product
+            try {
+                const data = await Product.create(newProduct)
+                res.status(201).json(data)
+            } catch (err) {
+                res.status(500)
+                throw new Error('unknowen server or DB error')
+            }
+        }
+    } else {
         res.status(401)
-        res.json(err)
+        throw new Error(`Unauthorized, no privilges`)
     }
 })
 
@@ -79,19 +80,24 @@ const addProduct = asyncHandler(async (req, res) => {
 // @route   DELETE /api/products/:id
 // @access  private
 const deleteProduct = asyncHandler(async (req, res) => {
-    const id = req.params.id
+    if (req.user.type === 'Admin' && req.user.status === 'Active') {
+        const id = req.params.id
 
-    // Check for product
-    const doc = await Product.findById(id)
+        // Check for product
+        const doc = await Product.findById(id)
 
-    if (doc) {
-        await Product.deleteOne({ _id: id })
-        res.status(201).json({
-            id: doc.id
-        })
+        if (doc) {
+            await Product.deleteOne({ _id: id })
+            res.status(200).json({
+                id: doc._id
+            })
+        } else {
+            res.status(400)
+            throw new Error('Invalid product id')
+        }
     } else {
-        res.status(400)
-        throw new Error('Invalid product id')
+        res.status(401)
+        throw new Error(`Unauthorized, no privilges`)
     }
 })
 
@@ -99,6 +105,10 @@ const deleteProduct = asyncHandler(async (req, res) => {
 // @route   PUT /api/products/:id
 // @access  private
 const editProduct = asyncHandler(async (req, res) => {
+    if (req.user.type === 'Admin' && req.user.status === 'Active') { } else {
+        res.status(401)
+        throw new Error(`Unauthorized, no privilges`)
+    }
     const { category, images, age, pieces, isFeatured, features, highlights, details, name, price, brand, tags } = req.body
     const id = req.params.id
